Remove dependency cycle on JsonInputField
