+++ packages/admin-panel/src/widgets/InputField/JsonEditor.js
