+++ packages/admin-panel/src/widgets/InputField/JsonInputField.js
