+++ packages/admin-panel/src/widgets/InputField/index.js
@@ -1,0 +1,5 @@
+import { registerInputFields } from './registerInputFields';
+
+export { InputField } from './InputField';
+
+registerInputFields();
